+++ src/builtin_functions/pow.cpp
@@ -2,14 +2,14 @@
 
 #include "circa.h"
 
-#include "math.h"
+#include <cmath>
 
 namespace circa {
 namespace pow_function {
 
     void evaluate(Term* caller)
     {
-        as_int(caller) = (int) pow(as_int(caller->input(0)), as_int(caller->input(1)));
+        as_int(caller) = (int) std::pow((float) as_int(caller->input(0)), as_int(caller->input(1)));
     }
 
     void setup(Branch& kernel)
